Refactor button file
