+++ src/__test__/App.test.js
@@ -1,14 +1,55 @@
 import React from 'react';
-
 import { create } from "react-test-renderer";
-import { EditorScreen } from "../components/screens/EditorScreen";
+import { MockedProvider } from '@apollo/react-testing';
+import gql from 'graphql-tag'
+import { LoadButton } from '../components/buttons/LoadButton';
 
-// test("Matches the snapshot", () => {
-//     const button = create(<EditorScreen />);
-//     expect(button.toJSON()).toMatchSnapshot();
-//   });
 
 it('CheckboxWithLabel changes the text after click', () => {
 
   expect(true).toBeTruthy();
+});
+
+
+const query = gql
+    `query GetGrabitByName($projectName: String!){
+                    allGrabits(nameProject: $projectName){
+                        edges{
+                            node{
+                                nameProject
+                            }
+                        }
+                    }
+				}`;
+
+const mocks = [
+  {
+    request: {
+      query: query,
+      variables: {
+        projectName: "prova8"
+      },
+    },
+    result: {
+      "data": {
+        "allGrabits": {
+          "edges": [
+            {
+              "node": {
+                "nameProject": "prova8"
+              }
+            }
+          ]
+        }
+      }
+    },
+  },
+];
+
+it('renders without error', () => {
+    create(
+        <MockedProvider mocks={mocks} addTypename={false}>
+          <LoadButton />
+        </MockedProvider>
+    );
 });
