+++ src/__test__/GrabitCreatorBtn.test.js
@@ -1,13 +1,13 @@
 import * as React from 'react';
 import { create } from "react-test-renderer";
 import { MockedProvider } from "@apollo/react-testing";
-import GrabitCreatorBtn from '../components/buttons/GrabitCreatorBtn';
+import CreateButton from '../components/buttons/CreateButton';
 
 describe('GrabitCreator Tests', () => {
   test("Matches the snapshot", () => {
     const button = create(
       <MockedProvider>
-        <GrabitCreatorBtn />
+        <CreateButton />
       </MockedProvider>
     );
     expect(button.toJSON()).toMatchSnapshot();
