+++ src/components/buttons/index.js
@@ -1,0 +1,4 @@
+export {default as CreateButton} from './CreateButton.js';
+export {default as DeleteButton} from './DeleteButton.js';
+export {default as LoadButton} from './LoadButton.js';
+export {default as SaveButton} from './SaveButton.js';
