+++ src/components/screens/EditorScreen.js
@@ -1,10 +1,7 @@
 import React from 'react';
 import { GraphEditor } from '../GraphEditor/GraphEditor';
 import { Container, Row, Col } from 'react-bootstrap';
-import SaveButton from '../SaveButton';
-import GrabitCreatorBtn from '../buttons/GrabitCreatorBtn';
-import DeleteButton from "../DeleteButton";
-import LoadButton from "../LoadButton";
+import { SaveButton, CreateButton, DeleteButton, LoadButton} from '../buttons';
 
 export class EditorScreen extends React.Component{
 
@@ -30,7 +27,7 @@ export class EditorScreen extends React.Component{
 							<LoadButton variant="outline-secondary" loadGraph={this.triggerLoadGraph} text="Carica"/>
 						</Row>
 						<Row className="mt-2 justify-content-center">
-							<GrabitCreatorBtn/>
+							<CreateButton/>
 						</Row>
 						<Row className="mt-2 justify-content-center">
 							<SaveButton trigger={this.triggerGetJSON} variant="outline-success" text="Salva"/>
